Remove some crates and rename others
