+++ .github/actions/desktop/action.yml
@@ -19,8 +19,8 @@ runs:
       run: sudo apt-get install -y libwayland-dev libxkbcommon-dev # Required for winit
     - name: Build
       shell: bash
-      run: cargo build -p maplibre-example
+      run: cargo build -p maplibre-demo
     - uses: actions/upload-artifact@v2
       with:
         name: maplibre-rs
-        path: target/x86_64-unknown-linux-gnu/debug/maplibre-example
+        path: target/x86_64-unknown-linux-gnu/debug/maplibre-demo
