+++ .gitignore
@@ -6,9 +6,6 @@ target/
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
-!maplibre-android/Cargo.lock
-!maplibre-apple/Cargo.lock
-!maplibre-example/Cargo.lock
 
 # These are backup files generated by rustfmt
 **/*.rs.bk
