+++ .idea/maplibre-rs.iml
@@ -5,17 +5,17 @@
       <sourceFolder url="file://$MODULE_DIR$/examples" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/libs/mbtiles/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/maplibre-android/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/maplibre-apple/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/maplibre-benchmark/benches" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/maplibre-benchmark/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/maplibre-build-tools/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/maplibre-core/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/maplibre-example/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/maplibre-style/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/maplibre-tilejson/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/maplibre-web/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/maplibre/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/android/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/apple/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/benchmarks/benches" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/benchmarks/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/maplibre-demo/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/web/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
     </content>
     <orderEntry type="inheritedJdk" />
