+++ .idea/runConfigurations/Run_demo__debug_enable_tracing_.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
-  <configuration default="false" name="Tesselation" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
-    <option name="command" value="bench --bench tessellation" />
+  <configuration default="false" name="Run demo (debug+enable-tracing)" type="CargoCommandRunConfiguration" factoryName="Cargo Command">
+    <option name="command" value="run -p maplibre-demo --features enable-tracing" />
     <option name="workingDirectory" value="file://$PROJECT_DIR$" />
     <option name="channel" value="DEFAULT" />
     <option name="requiredFeatures" value="true" />
