+++ android/Cargo.toml
@@ -1,9 +1,10 @@
 [package]
-name = "maplibre-android"
+name = "android"
 version = "0.1.0"
 description = ""
 categories = []
 edition = "2021"
+publish = false
 
 [dependencies]
 maplibre = { path = "../maplibre" }
@@ -13,6 +14,7 @@ env_logger = "0.9"
 ndk-glue = "0.5.0" # version is required by winit
 
 [lib]
+name = "maplibre_android"
 crate-type = ["rlib", "cdylib"]
 
 [package.metadata.android]
