+++ apple/src/lib.rs
@@ -4,7 +4,7 @@ pub use std::time::Instant;
 
 // TODO clippy
 // #[cfg(not(any(target_os = "macos", target_os = "ios")))]
-// compile_error!("maplibre-apple works only on macOS and iOS.");
+// compile_error!("apple works only on macOS and iOS.");
 
 #[no_mangle]
 pub fn maplibre_apple_main() {
