+++ maplibre-core/src/style/mod.rs
@@ -2,6 +2,6 @@ pub mod layer;
 pub mod source;
 mod style;
 
-pub use crate::style::*;
+pub use style::*;
 
 pub use cint::*;
