+++ maplibre-core/src/style/source.rs
