+++ maplibre-core/src/tilejson/mod.rs
@@ -1,0 +1,3 @@
+mod tilejson;
+
+pub use tilejson::*;
