+++ maplibre-core/style-spec-v8.json
