+++ maplibre-demo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "maplibre-example"
+name = "maplibre-demo"
 version = "0.1.0"
 description = ""
 categories = []
