--- maplibre-style/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "maplibre-style"
-version = "0.1.0"
-description = "A library for decoding style specs"
-readme = "README.md"
-categories = ["encoding"]
-edition = "2021"
-build = "build.rs"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-csscolorparser = { version = "0.5", features = ["serde", "cint"]}
-cint = "0.2"
-
-[build-dependencies]
-serde_json = "1.0"
-
-[dev-dependencies]
-serde_json = "1.0"
