--- maplibre-tilejson/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "maplibre-tilejson"
-version = "0.1.0"
-description = "A library for decoding and encoding tilejson specs"
-categories = []
-edition = "2021"
-
-[dependencies]
-serde = { version = "1.0", features = ["derive"] }
-
-[dev-dependencies]
-serde_json = "1.0"
