--- maplibre-tilejson/src/lib.rs
@@ -1,3 +1,0 @@
-mod tilejson;
-
-pub use crate::tilejson::*;
