+++ web/src/lib.rs
@@ -2,4 +2,4 @@ pub use maplibre_core::*;
 
 // TODO clippy
 // #[cfg(not(target_arch = "wasm32"))]
-// compile_error!("maplibre-web works only on wasm32.");
+// compile_error!("web works only on wasm32.");
