+++ web/web/webpack.config.js
